docs(ops): production operations reference; TimescaleDB 2.28.2 pins; README sync
Adds docs/operations.md covering the TimescaleDB upgrade runbook (single
ALTER EXTENSION multi-minor jumps, the 2.27+ composite-bloom catalog
rename, never landing on 2.28.0/.1), the production database checklist
(chunk-skipping GUC required before first ingest, instance sizing and
retention/compression flag values), alerting guidance (lag and restarts
are the pager signals; retry-exhaustion is a post-mortem breadcrumb;
cursor_missing is now alertable), and the connection-topology constraint
for ever fronting ingest with a transaction-pooling proxy.

TimescaleDB image pins move to 2.28.2 everywhere (docker-compose, the
CNPG image build, both build workflows, and CI's test service).

The GraphQL README drops a documented-but-nonexistent mutation and error
code, references the complexity limit by its flag instead of a stale
literal, adds the Blend queries, and documents the API behavior
introduced by this hardening: since/until with the 7-day root default,
uniform 100-per-page caps, the depth limit, gated introspection, error
masking, and the request timeout. Ledger cursor casts widen to bigint.

+++ internal/data/ingest_store.go
@@ -198,7 +198,7 @@ func (m *IngestStoreModel) CompareAndSwap(ctx context.Context, dbTx pgx.Tx, curs
 func (m *IngestStoreModel) UpdateMin(ctx context.Context, dbTx pgx.Tx, cursorName string, ledger uint32) error {
 	const query = `
 		UPDATE ingest_store
-		SET value = LEAST(value::integer, $2)::text
+		SET value = LEAST(value::bigint, $2)::text
 		WHERE key = $1
 	`
 	start := time.Now()
